UI Test (complete)
Start menu,
Settings menu,
Pause menu    and
Billing Menu.
How to use:
It's a good idea to check the UI Programming in the notion Project Wiki before copying, debugging, and merging canvas with all built-in menu functions.
If you do not understand or need to add, please ask me directly.